experiments/requirements.txt
llama-index==0.14.13
numpy==2.4.1
llama-index-retrievers-bm25==0.6.5
llama-index-postprocessor-cohere-rerank==0.5.1
llama-index-embeddings-huggingface==0.6.1
llama-index-llms-ollama==0.9.1
datasets==4.5.0
ipywidgets==8.1.8
jupyter==1.1.1
langchain-groq==0.3.8
langchain_community==0.3.29
langchain==0.3.27
langchain-core==0.3.83
openpyxl==3.1.5
pgvector==0.4.2 
psycopg2-binary==2.9.11
llama-index-vector-stores-postgres==0.7.3 
google-auth==2.49.0.dev0
google-auth-oauthlib==1.3.0
google-api-python-client==2.190.0